Revamped settings to include baby picker and option to delete baby. Added preferences for kg and lb.
