+++ Feedbridge/Resources/Localizable.xcstrings
@@ -1,6 +1,16 @@
 {
   "sourceLanguage" : "en",
   "strings" : {
+    "%.2f %@" : {
+      "localizations" : {
+        "en" : {
+          "stringUnit" : {
+            "state" : "new",
+            "value" : "%1$.2f %2$@"
+          }
+        }
+      }
+    },
     "%.2f lb" : {
 
     },
@@ -65,9 +75,6 @@
     "Add Baby" : {
 
     },
-    "Add Data" : {
-
-    },
     "Add Dehydration Check" : {
 
     },
@@ -111,7 +118,7 @@
     "Amount: %lldml" : {
 
     },
-    "Baby Debug View" : {
+    "Are you sure you want to delete this baby?" : {
 
     },
     "Baby icon" : {
@@ -238,6 +245,12 @@
     "Dehydration Symptoms" : {
 
     },
+    "Delete" : {
+
+    },
+    "Delete Baby" : {
+
+    },
     "Diaper #" : {
 
     },
@@ -262,10 +275,10 @@
     "Error" : {
 
     },
-    "Feed Details" : {
+    "Feed #" : {
 
     },
-    "Feed #" : {
+    "Feed Details" : {
 
     },
     "Feed Entries" : {
@@ -514,9 +527,6 @@
     "No data added" : {
 
     },
-    "No weight entries" : {
-
-    },
     "NOTIFICATION_PERMISSIONS_DESCRIPTION" : {
       "localizations" : {
         "en" : {
@@ -603,10 +613,16 @@
         }
       }
     },
+    "Select baby" : {
+
+    },
     "Select Date & Time" : {
 
     },
     "Selected" : {
+
+    },
+    "Settings" : {
 
     },
     "Social Support Questionnaire" : {
@@ -650,10 +666,10 @@
         }
       }
     },
-    "Stool Details" : {
+    "Stool #" : {
 
     },
-    "Stool #" : {
+    "Stool Details" : {
 
     },
     "Stool Drop" : {
@@ -728,11 +744,14 @@
         }
       }
     },
+    "Use Kilograms" : {
+
+    },
     "Void Details" : {
 
     },
     "Voids" : {
-    
+
     },
     "Volume" : {
