+++ Feedbridge/Views/DashboardView.swift
@@ -7,8 +7,7 @@ struct DashboardView: View {
     @Environment(FeedbridgeStandard.self) private var standard
     @Binding var presentingAccount: Bool
     
-    @State private var babies: [Baby] = []
-    @State private var selectedBabyId: String?
+    @AppStorage(UserDefaults.selectedBabyIdKey) private var selectedBabyId: String?
     @State private var isLoading = true
     @State private var errorMessage: String?
     @State private var baby: Baby?
@@ -32,7 +31,6 @@ struct DashboardView: View {
                 }
             }
             .task {
-                await loadBabies()
                 await loadBaby()
             }
         }
@@ -41,7 +39,6 @@ struct DashboardView: View {
     @ViewBuilder private var mainContent: some View {
         ScrollView {
             VStack(spacing: 16) {
-                babyPicker
                 if let baby {
                     WeightsSummaryView(entries: baby.weightEntries.weightEntries)
                     FeedsSummaryView(entries: baby.feedEntries.feedEntries)
@@ -52,69 +49,6 @@ struct DashboardView: View {
             .padding()
         }
     }
-    
-    @ViewBuilder private var babyPicker: some View {
-        Menu {
-            ForEach(babies) { baby in
-                Button {
-                    selectedBabyId = baby.id
-                    UserDefaults.standard.selectedBabyId = baby.id
-                } label: {
-                    HStack {
-                        Text(baby.name)
-                        Spacer()
-                        if baby.id == selectedBabyId {
-                            Image(systemName: "checkmark")
-                                .accessibilityLabel("Selected")
-                        }
-                    }
-                }
-            }
-            Divider()
-            NavigationLink("Add New Baby") {
-                AddSingleBabyView(onSave: {
-                    Task {
-                        await loadBabies()
-                    }
-                })
-            }
-        } label: {
-            HStack {
-                Image(systemName: "person.crop.circle")
-                    .accessibilityLabel("Baby icon")
-                Text(babies.first(where: { $0.id == selectedBabyId })?.name ?? "Select Baby")
-                Image(systemName: "chevron.down")
-                    .accessibilityLabel("Menu dropdown")
-            }
-            .foregroundColor(.primary)
-            .padding()
-            .frame(maxWidth: .infinity)
-            .background(Color(.systemBackground))
-            .cornerRadius(8)
-            .shadow(radius: 2)
-        }
-    }
-    
-    private func loadBabies() async {
-        isLoading = true
-        errorMessage = nil
-        
-        do {
-            babies = try await standard.getBabies()
-            if let savedId = UserDefaults.standard.selectedBabyId,
-               babies.contains(where: { $0.id == savedId }) {
-                selectedBabyId = savedId
-            } else {
-                selectedBabyId = babies.first?.id
-                UserDefaults.standard.selectedBabyId = selectedBabyId
-            }
-        } catch {
-            errorMessage = "Failed to load babies: \(error.localizedDescription)"
-        }
-        
-        isLoading = false
-    }
-
     private func loadBaby() async {
         guard let babyId = selectedBabyId else {
             baby = nil
